BUGFIX: Respect both, "operator" and "keepAllOptionsOnSelection" when creating facets.
When keepAllOptionsOnSelection=1 and operator=and, this effectively only allows
a single-select mode. Every prediction is wrong and there's even the chance of
hitting a dead end when adding a new option to the current list of restrictions
that isn't even allowed.

When keepAllOptionsOnSelection=0, predictions will be correct. So we can use this
for multi-select modes no matter if the operator is set to "and" or "or".

When keepAllOptionsOnSelection=1 and operator is set to "or", prediction might be
wrong (so "count" is off), but adding a new value to the list of selected options
the worst thing that can happen is not changing the result set.